requirements.txt
RPi.GPIO
mfrc522==0.0.7
RPLCD==1.3.1
python-mpd2==3.1.0
Flask
flask_sqlalchemy  # Optional, entfernen Sie diese Zeile, wenn Sie SQLAlchemy nicht verwenden möchten
